requirements.txt
ollama>=0.2.0
python-docx>=1.1.0
PyPDF2>=3.0.0
pillow>=10.0.0
python-magic>=0.4.27
openpyxl>=3.1.0
pandas>=2.0.0
chardet>=5.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
pytesseract>=0.3.10
opencv-python>=4.8.0
